+++ CHANGELOG.md
@@ -3,7 +3,17 @@ arcgis cookbook CHANGELOG
 
 This file is used to list changes made in each version of the arcgis cookbook.
 
-1.1.0
+1.1.1
 -----
 - Initial release of arcgis cookbook
+
+1.1.2
+----
+- Support configuration of highly available Portal for ArcGIS
+- Fixed problem with NetFx3 windows feature installation on Windows 8.1
+- Fixed problem with missing '\items\portal\properties.json' file
+- Use path relative to 'install_dir' path for server and portal software authorization tools on linux
+- Improved 'desktop' and 'licensemanager' recipes
+- Supported domain user accounts
+